refactor(tests): reorg and dedup all IT
