+++ eslint.config.mjs
@@ -138,6 +138,7 @@ export default [
       'jest/no-disabled-tests': 'warn',
       'jest/no-focused-tests': 'error',
       'jest/no-identical-title': 'error',
+      '@typescript-eslint/explicit-function-return-type': 'off',
       'jest/prefer-to-have-length': 'warn',
       'jest/valid-expect': 'error',
       '@typescript-eslint/no-explicit-any': 'off',
