--- tests/fixtures/folder/1.txt
@@ -1,1 +1,0 @@
-Hello World!
