--- tests/fixtures/nested/extra/extra.txt
@@ -1,1 +1,0 @@
-Extra content.
