--- tests/fixtures/nested/nested.txt
@@ -1,1 +1,0 @@
-I am nested.
