+++ tests/unit/abort.spec.ts
@@ -1,14 +1,8 @@
 import { BatchId, Bee, MantarayNode, RedundancyLevel, Topic } from '@ethersphere/bee-js';
 
-import { DEFAULT_MOCK_SIGNER, DUMMY_BATCH_ID } from '../utils';
+import { createInitializedFileManager, DEFAULT_MOCK_SIGNER, DUMMY_BATCH_ID } from '../utils';
 
-import {
-  applyDefaultMocks,
-  createInitializedFileManager,
-  createMockDriveInfo,
-  createMockNodeAddresses,
-  seedRecords,
-} from './mock';
+import { applyDefaultMocks, createMockDriveInfo, createMockNodeAddresses, seedRecords } from './mock';
 
 import { FileRecord, ListDepth, NodeType } from '@/types';
 import { DriveError } from '@/utils';
